refactor: 불필요한 폴더 삭제 및 불필요한 상수 삭제를 통한 구조 변경
