+++ app/(admin)/admin/layout.tsx
@@ -1,9 +1,7 @@
 import { ReactNode } from 'react';
 import type { Metadata } from 'next';
-import { RTProvider, StyledProvider } from '@/lib/Provider';
+import { RTProvider, StyledProvider, DM_SANS } from '@/lib';
 import ScrollToTop from '@/components/Common/ScrollTop';
-import { DM_SANS } from '@/app/fonts';
-import { NextAuthProvider } from '@/app/lib/Provider/SessionProvider';
 
 export const metadata: Metadata = {
   title: '관리자',
@@ -15,10 +13,8 @@ const AdminLayout = ({ children }: { children: ReactNode }) => {
       <body className={DM_SANS.className}>
         <StyledProvider>
           <RTProvider>
-            <NextAuthProvider>
-              <ScrollToTop />
-              {children}
-            </NextAuthProvider>
+            <ScrollToTop />
+            {children}
           </RTProvider>
         </StyledProvider>
       </body>
