--- app/fonts/index.ts
@@ -1,3 +1,0 @@
-import { DM_Sans as dmSans } from 'next/font/google';
-
-export const DM_SANS = dmSans({ weight: ['400', '500', '700'], subsets: ['latin'], display: 'swap' });
